Integration fin

+++ src/main/java/com/safetynet/web/controller/MedicalRecordsController.java
@@ -1,16 +1,17 @@
 package com.safetynet.web.controller;
 
 import com.safetynet.dto.MedicalRecordsDTO;
+import com.safetynet.model.Medicalrecords;
 import com.safetynet.service.IMedicalRecordsService;
+import com.safetynet.web.exceptions.BadArgumentsException;
+import com.safetynet.web.exceptions.FirestationNotFoundException;
 import com.safetynet.web.exceptions.MedicalrecordsNotFoundException;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import java.net.URI;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
 
@@ -27,51 +28,51 @@ public class MedicalRecordsController {
     }
 
     @PostMapping(value = "medicalRecord")
-    public ResponseEntity<Void> addMedicalRecord(@RequestBody MedicalRecordsDTO medicalrecords) {
-        log.info("Add medicalrecords : " + medicalrecords);
+    public ResponseEntity<MedicalRecordsDTO> addMedicalRecord(@RequestBody MedicalRecordsDTO medicalrecords) {
+        log.debug("POST : create MedicalRecords : {}", medicalrecords);
         AtomicReference<ResponseEntity> rep = new AtomicReference<>();
-        medicalRecordsService.add(medicalrecords)
-                .ifPresentOrElse(retour ->
-                {
-                    URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{prenom}/{nom}")
-                            .buildAndExpand(medicalrecords.getFirstName(), medicalrecords.getLastName())
-                            .toUri();
-                    rep.set(ResponseEntity.created(location).build());
-                }, () ->
-                {
-                    rep.set(ResponseEntity.noContent().build());
-                });
 
+        if (medicalrecords.getFirstName().isBlank() || medicalrecords.getLastName().isBlank()) {
+            log.error("POST : firstName or lastName are not null for create medical records");
+            throw new BadArgumentsException("POST : firstName or lastName are not null for create medical records");
+        }
+        medicalRecordsService.add(medicalrecords)
+                .ifPresent(retour ->
+                {
+                    log.info("POST: create medicalRecords - SUCCESS");
+                    rep.set(ResponseEntity.status(HttpStatus.CREATED).body(retour));
+                });
         return rep.get();
     }
 
     @PutMapping(value = "medicalRecord")
-    public ResponseEntity<Void> updateMedicalrecord(@RequestBody MedicalRecordsDTO medicalRecord) {
-        log.info("Update medicalRecord : " + medicalRecord);
-        AtomicReference<ResponseEntity> rep = new AtomicReference<>();
-        medicalRecordsService.update(medicalRecord)
-                .ifPresentOrElse(retour ->
-                {
-                    URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{nom}/{prenom}")
-                            .buildAndExpand(medicalRecord.getLastName(), medicalRecord.getFirstName())
-                            .toUri();
-                    rep.set(ResponseEntity.created(location).build());
-                }, () ->
-                {
-                    rep.set(ResponseEntity.noContent().build());
-                });
+    public ResponseEntity<MedicalRecordsDTO> updateMedicalrecord(@RequestBody MedicalRecordsDTO medicalRecord) {
+        log.debug("PUT : update medical record : " + medicalRecord);
 
-        return rep.get();
+        if (medicalRecord.getFirstName().isBlank() || medicalRecord.getLastName().isBlank()) {
+            log.error("PUT : firstName or lastName are not null for update medical records");
+            throw new BadArgumentsException("PUT : firstName or lastName are not null for update medical records");
+        }
+        return ResponseEntity.status(HttpStatus.OK.value())
+                .body(medicalRecordsService.update(medicalRecord)
+                .orElseThrow(MedicalrecordsNotFoundException::new));
     }
 
     @DeleteMapping(value = "medicalRecord")
     public ResponseEntity<Void> deletePerson(@RequestBody MedicalRecordsDTO medicalRecord) {
-        log.info("Delete medicalRecord :  " + medicalRecord);
+        log.debug("DELETE : medicalRecord :  {}", medicalRecord);
+
+        if (medicalRecord.getFirstName().isBlank() || medicalRecord.getLastName().isBlank()) {
+            log.error("DELETE : firstName or lastName are not null for update medical records");
+            throw new BadArgumentsException("DELETE : firstName or lastName are not null for update medical records");
+        }
+
         Boolean retour = medicalRecordsService.delete(medicalRecord);
         if (!retour) {
-            throw new MedicalrecordsNotFoundException("La personne se nommant " + medicalRecord.getLastName() + "est introuvable.");
+            throw new MedicalrecordsNotFoundException("The person naming himself " + medicalRecord.getLastName() + "is not found.");
         }
-        return new ResponseEntity<>(HttpStatus.OK);
+        log.info("Delete :of the medical file is carried out");
+        return new ResponseEntity<>(HttpStatus.ACCEPTED);
     }
 
 }
